Rename solidmd to markdown
