+++ docs/src/components/sdk-support-table/sdk-support-table.tsx
@@ -1,7 +1,7 @@
 //@ts-ignore
 import style from './sdk-support-table.module.scss';
 import entries from 'object.entries';
-import {SolidMd} from '~/components/SolidMd.jsx';
+import {SolidMd} from '~/components/markdown.jsx';
 
 const SDKSupportTable = (props: { supportItems?: Object }) => {
     function support(support:any, sdk:any) {
