+++ docs/src/routes/light.tsx
@@ -1,4 +1,4 @@
-import {SolidMd} from '~/components/SolidMd';
+import {SolidMd} from '~/components/markdown';
 import ref from '../reference/latest';
 import {Items} from '../components/style-spec/specs/items';
 function Light() {
